# agents/openapi.yaml
openapi: 3.1.0
info:
  title: startupShop Agent API
  version: 0.4.0
  description: Agent-navigable API for startup listings and offer submissions.
servers:
  - url: http://localhost:3000
paths:
...
  /api/offers:
    post:
      summary: Submit offer
      description: Validates offer payload against `schemas/offer.schema.json` and stores it in SQLite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfferRequest'
      responses:
        '201':
          description: Offer accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/OfferCreated'
        '400':
          description: Invalid payload or unknown startup id
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
components:
  schemas:
...
    OfferRequest:
      type: object
      required: [startup_id, buyer_name, buyer_email, offer_amount_usd, message]
      properties:
        startup_id:
          type: string
        buyer_name:
          type: string
        buyer_email:
          type: string
          format: email
        offer_amount_usd:
          type: number
        message:
          type: string
    OfferCreated:
      type: object
      required: [id, created_at, startup_id]
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
        startup_id:
          type: string